make util go away
