+++ pinot-query-planner/src/test/java/org/apache/pinot/query/QueryEnvironmentTestBase.java
@@ -20,8 +20,21 @@ package org.apache.pinot.query;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
+import org.apache.calcite.jdbc.CalciteSchemaBuilder;
+import org.apache.pinot.common.config.provider.TableCache;
+import org.apache.pinot.core.routing.RoutingManager;
+import org.apache.pinot.query.catalog.PinotCatalog;
+import org.apache.pinot.query.routing.WorkerManager;
+import org.apache.pinot.query.testutils.MockRoutingManagerFactory;
+import org.apache.pinot.query.type.TypeFactory;
+import org.apache.pinot.query.type.TypeSystem;
+import org.apache.pinot.spi.data.FieldSpec;
+import org.apache.pinot.spi.data.Schema;
+import org.apache.pinot.spi.data.readers.GenericRow;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
 
@@ -33,12 +46,25 @@ public class QueryEnvironmentTestBase {
   public static final Map<String, List<String>> SERVER2_SEGMENTS =
       ImmutableMap.of("a_REALTIME", ImmutableList.of("a3"), "c_OFFLINE", ImmutableList.of("c2", "c3"),
           "d_REALTIME", ImmutableList.of("d2"), "d_OFFLINE", ImmutableList.of("d3"));
+  public static final Schema.SchemaBuilder SCHEMA_BUILDER;
+  public static final int NUM_ROWS = 5;
+  public static final String[] STRING_FIELD_LIST = new String[]{"foo", "bar", "alice", "bob", "charlie"};
+  public static final int[] INT_FIELD_LIST = new int[]{1, 42};
+  static {
+    SCHEMA_BUILDER = new Schema.SchemaBuilder()
+        .addSingleValueDimension("col1", FieldSpec.DataType.STRING, "")
+        .addSingleValueDimension("col2", FieldSpec.DataType.STRING, "")
+        .addDateTime("ts", FieldSpec.DataType.LONG, "1:MILLISECONDS:EPOCH", "1:HOURS")
+        .addMetric("col3", FieldSpec.DataType.INT, 0)
+        .setSchemaName("defaultSchemaName");
+  }
+
   protected QueryEnvironment _queryEnvironment;
 
   @BeforeClass
   public void setUp() {
     // the port doesn't matter as we are not actually making a server call.
-    _queryEnvironment = QueryEnvironmentTestUtils.getQueryEnvironment(3, 1, 2, SERVER1_SEGMENTS, SERVER2_SEGMENTS);
+    _queryEnvironment = getQueryEnvironment(3, 1, 2, SERVER1_SEGMENTS, SERVER2_SEGMENTS);
   }
 
   @DataProvider(name = "testQueryDataProvider")
@@ -71,5 +97,43 @@ public class QueryEnvironmentTestBase {
         new Object[]{"SELECT a.col1, b.col2 FROM a JOIN b ON a.col1 = b.col1 WHERE a.col2 IN ('foo', 'bar') AND"
             + " b.col2 NOT IN ('alice', 'charlie')"},
     };
+  }
+
+  public static List<GenericRow> buildRows(String tableName) {
+    List<GenericRow> rows = new ArrayList<>(NUM_ROWS);
+    for (int i = 0; i < NUM_ROWS; i++) {
+      GenericRow row = new GenericRow();
+      row.putValue("col1", STRING_FIELD_LIST[i % STRING_FIELD_LIST.length]);
+      row.putValue("col2", STRING_FIELD_LIST[i % (STRING_FIELD_LIST.length - 2)]);
+      row.putValue("col3", INT_FIELD_LIST[i % INT_FIELD_LIST.length]);
+      row.putValue("ts", tableName.endsWith("_O")
+          ? System.currentTimeMillis() - TimeUnit.DAYS.toMillis(2) : System.currentTimeMillis());
+      rows.add(row);
+    }
+    return rows;
+  }
+
+  public static QueryEnvironment getQueryEnvironment(int reducerPort, int port1, int port2,
+      Map<String, List<String>> segmentMap1, Map<String, List<String>> segmentMap2) {
+    MockRoutingManagerFactory factory = new MockRoutingManagerFactory(port1, port2)
+        .registerTable(SCHEMA_BUILDER.setSchemaName("a").build(), "a_REALTIME")
+        .registerTable(SCHEMA_BUILDER.setSchemaName("b").build(), "b_REALTIME")
+        .registerTable(SCHEMA_BUILDER.setSchemaName("c").build(), "c_OFFLINE")
+        .registerTable(SCHEMA_BUILDER.setSchemaName("d").build(), "d");
+    for (Map.Entry<String, List<String>> entry : segmentMap1.entrySet()) {
+      for (String segment : entry.getValue()) {
+        factory.registerSegment(port1, entry.getKey(), segment);
+      }
+    }
+    for (Map.Entry<String, List<String>> entry : segmentMap2.entrySet()) {
+      for (String segment : entry.getValue()) {
+        factory.registerSegment(port2, entry.getKey(), segment);
+      }
+    }
+    RoutingManager routingManager = factory.buildRoutingManager();
+    TableCache tableCache = factory.buildTableCache();
+    return new QueryEnvironment(new TypeFactory(new TypeSystem()),
+        CalciteSchemaBuilder.asRootSchema(new PinotCatalog(tableCache)),
+        new WorkerManager("localhost", reducerPort, routingManager));
   }
 }
