+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/QueryServerEnclosure.java
@@ -33,6 +33,7 @@ import org.apache.pinot.query.runtime.QueryRunner;
 import org.apache.pinot.query.runtime.plan.DistributedStagePlan;
 import org.apache.pinot.query.service.QueryConfig;
 import org.apache.pinot.query.testutils.MockInstanceDataManagerFactory;
+import org.apache.pinot.query.testutils.QueryTestUtils;
 import org.apache.pinot.spi.data.Schema;
 import org.apache.pinot.spi.env.PinotConfiguration;
 import org.apache.pinot.spi.utils.builder.TableNameBuilder;
@@ -74,7 +75,7 @@ public class QueryServerEnclosure {
     try {
       _instanceDataManager = factory.buildInstanceDataManager();
       _helixManager = mockHelixManager(factory.buildSchemaMap());
-      _queryRunnerPort = QueryEnvironmentTestUtils.getAvailablePort();
+      _queryRunnerPort = QueryTestUtils.getAvailablePort();
       _runnerConfig.put(QueryConfig.KEY_OF_QUERY_RUNNER_PORT, _queryRunnerPort);
       _runnerConfig.put(QueryConfig.KEY_OF_QUERY_RUNNER_HOSTNAME,
           String.format("Server_%s", QueryConfig.DEFAULT_QUERY_RUNNER_HOSTNAME));
