+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/service/QueryDispatcherTest.java
@@ -25,11 +25,11 @@ import java.util.Map;
 import java.util.Random;
 import org.apache.pinot.query.QueryEnvironment;
 import org.apache.pinot.query.QueryEnvironmentTestBase;
-import org.apache.pinot.query.QueryEnvironmentTestUtils;
 import org.apache.pinot.query.QueryTestSet;
 import org.apache.pinot.query.planner.PlannerUtils;
 import org.apache.pinot.query.planner.QueryPlan;
 import org.apache.pinot.query.runtime.QueryRunner;
+import org.apache.pinot.query.testutils.QueryTestUtils;
 import org.mockito.Mockito;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -50,7 +50,7 @@ public class QueryDispatcherTest extends QueryTestSet {
       throws Exception {
 
     for (int i = 0; i < QUERY_SERVER_COUNT; i++) {
-      int availablePort = QueryEnvironmentTestUtils.getAvailablePort();
+      int availablePort = QueryTestUtils.getAvailablePort();
       QueryRunner queryRunner = Mockito.mock(QueryRunner.class);
       QueryServer queryServer = new QueryServer(availablePort, queryRunner);
       queryServer.start();
@@ -61,7 +61,7 @@ public class QueryDispatcherTest extends QueryTestSet {
     List<Integer> portList = Lists.newArrayList(_queryServerMap.keySet());
 
     // reducer port doesn't matter, we are testing the worker instance not GRPC.
-    _queryEnvironment = QueryEnvironmentTestUtils.getQueryEnvironment(1, portList.get(0), portList.get(1),
+    _queryEnvironment = QueryTestUtils.getQueryEnvironment(1, portList.get(0), portList.get(1),
         QueryEnvironmentTestBase.SERVER1_SEGMENTS, QueryEnvironmentTestBase.SERVER2_SEGMENTS);
   }
